LPD-48888 Create and apply Base test class to reduce code duplication.
